inventory subsystem: goes from beehive to inventory and vice versa

+++ Source/Honeycomber/Public/Utils/Structs.h
@@ -63,5 +63,5 @@ struct FBeeFrameData
 	GENERATED_BODY()
 
 	UPROPERTY()
-		float HoneyFill;
+		uint8 HoneyFill;
 };
